// client/src/store/auth/constants.ts
export enum AuthTypes {
    SIGNIN_REQUEST = 'SIGNIN_REQUEST',
    SIGNIN_SUCCESS = 'SIGNIN_SUCCESS',
    SIGNIN_FAILURE = 'SIGNIN_FAILURE',

    SIGNUP_REQUEST = 'SIGNUP_REQUEST',
    SIGNUP_SUCCESS = 'SIGNUP_SUCCESS',
    SIGNUP_FAILURE = 'SIGNUP_FAILURE',

    SIGNOUT_REQUEST = 'SIGNOUT_REQUEST',
    SIGNOUT_SUCCESS = 'SIGNOUT_SUCCESS',
    SIGNOUT_FAILURE = 'SIGNOUT_FAILURE',

    CHECK_USER_REQUEST = 'CHECK_USER_REQUEST',
    CHECK_USER_SUCCESS = 'CHECK_USER_SUCCESS',
    CHECK_USER_FAILURE = 'CHECK_USER_FAILURE',

    UPLOAD_USER_IMAGE_REQUEST = 'UPLOAD_USER_IMAGE_REQUEST',
    UPLOAD_USER_IMAGE_SUCCESS = 'UPLOAD_USER_IMAGE_SUCCESS',
    UPLOAD_USER_IMAGE_FAILURE = 'UPLOAD_USER_IMAGE_FAILURE',
}